Changed WorkshopBaseDto class.
