+++ OutOfSchool/OutOfSchool.BusinessLogic/Models/Workshops/WorkshopBaseDto.cs
@@ -1,9 +1,36 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using OutOfSchool.BusinessLogic.Models.Workshops.IncompletedWorkshops;
+using OutOfSchool.Services.Enums;
 
 namespace OutOfSchool.BusinessLogic.Models.Workshops;
 
 public class WorkshopBaseDto : WorkshopWithTeachersDto, IValidatableObject
 {
     public Guid Id { get; set; }
+
+    public IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
+    {
+        // TODO: Validate DateTimeRanges are not empty when frontend is ready
+        foreach (var dateTimeRange in DateTimeRanges)
+        {
+            if (dateTimeRange.StartTime > dateTimeRange.EndTime)
+            {
+                yield return new ValidationResult(
+                    "End date can't be earlier that start date");
+            }
+
+            if (dateTimeRange.Workdays.IsNullOrEmpty() || dateTimeRange.Workdays.Any(workday => workday == DaysBitMask.None))
+            {
+                yield return new ValidationResult(
+                    "Workdays are required");
+            }
+
+            var daysHs = new HashSet<DaysBitMask>();
+            if (!dateTimeRange.Workdays.All(daysHs.Add))
+            {
+                yield return new ValidationResult(
+                    "Workdays contain duplications");
+            }
+        }
+    }
 }
